Fix Project Test
- Using enum value (instead of label) for Filter/POST/PUT for Project
- Fix zero project query bug
- Add project_country and reporting_ns multiple filter by ID
